Moved badge_group into a folder
