+++ src/components/badge/_index.scss
@@ -1,4 +1,4 @@
 @import 'badge';
-@import 'badge_group';
+@import 'badge_group/index';
 @import 'beta_badge/index';
 @import 'notification_badge/index';
