+++ src/components/badge/badge_group/_badge_group.scss
