+++ src/components/badge/badge_group/_index.scss
@@ -1,0 +1,1 @@
+@import './badge_group';
