+++ src/components/badge/badge_group/index.ts
@@ -1,0 +1,1 @@
+export { EuiBadgeGroup } from './badge_group';
